# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Access Monitor API",
    "version": "1.0.0",
    "description": "JWT login for dashboard reads; per-device API key for Pi writes. All timestamps are server-generated. Import this file into Swagger Editor or Postman. Cookies are also accepted for dashboard reads."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "deviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "smart_access_token"
      }
    }
  },
  "paths": {
    "/api/login": {
      "post": {
        "summary": "Log in (also sets an HTTP-only cookie; token expires in 8 hours)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "token": "<JWT>"
                }
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "username": "admin",
                "password": "<your-password>"
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard": {
      "get": {
        "summary": "Current readings (aliases of the same handler)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "temperature": 24.6,
                  "humidity": 58,
                  "gas_status": "safe",
                  "fan_status": "off",
                  "device_online": true,
                  "last_seen": "2026-10-03T06:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad input",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Invalid request"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Login required"
                }
              }
            }
          },
          "500": {
            "description": "Server/database failure; retry later",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Internal server error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown device",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Device not found"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "device_uid",
            "in": "query",
            "description": "Filter by registered device UID. Omit to query all devices (latest device for current KPI).",
            "schema": {
              "type": "string"
            },
            "example": "main-entrance-pi"
          }
        ],
        "description": "Online means a valid reading was received within 2 minutes. Before first telemetry, readings/status strings and last_seen are null, online is false."
      }
    },
...
  }
}